chạy được thêm discount

+++ DBMS_NoiThat/Entity/Discount.cs
@@ -8,15 +8,15 @@ namespace DBMS_NoiThat.Entity
 {
     public class Discount
     {
-        string MaGiamGia ;
-        int MaSanPham ;
-        int MaKhachHang ;
-        float SoLuongGiam ;
-        DateTime NgayApDung ;
+        string MaGiamGia;
+        int MaSanPham;
+        int MaKhachHang;
+        decimal SoLuongGiam;
+        DateTime NgayApDung;
         DateTime NgayKetThuc;
         string LiDo;
 
-        public Discount(string maGiamGia, int maSanPham, int maKhachHang, float soLuongGiam, DateTime ngayApDung, DateTime ngayKetThuc, string liDo)
+        public Discount(string maGiamGia, int maSanPham, int maKhachHang, decimal soLuongGiam, DateTime ngayApDung, DateTime ngayKetThuc, string liDo)
         {
             MaGiamGia = maGiamGia;
             MaSanPham = maSanPham;
@@ -30,7 +30,7 @@ namespace DBMS_NoiThat.Entity
         public string MaGiamGia1 { get => MaGiamGia; set => MaGiamGia = value; }
         public int MaSanPham1 { get => MaSanPham; set => MaSanPham = value; }
         public int MaKhachHang1 { get => MaKhachHang; set => MaKhachHang = value; }
-        public float SoLuongGiam1 { get => SoLuongGiam; set => SoLuongGiam = value; }
+        public decimal SoLuongGiam1 { get => SoLuongGiam; set => SoLuongGiam = value; }
         public DateTime NgayApDung1 { get => NgayApDung; set => NgayApDung = value; }
         public DateTime NgayKetThuc1 { get => NgayKetThuc; set => NgayKetThuc = value; }
         public string LiDo1 { get => LiDo; set => LiDo = value; }
